Move notifications and pages to `users` app
