--- apps/bases/jest.config.js
@@ -1,17 +1,0 @@
-module.exports = {
-  displayName: 'bases',
-  preset: '../../jest.preset.js',
-  setupFiles: ['../../jest.setup.ts'],
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/bases',
-  collectCoverage: true,
-};
