+++ apps/users/src/app/dtos/index.ts
@@ -23,6 +23,7 @@
 
 export * from './blocking.dto';
 export * from './content.dto';
+export * from './delete-page.dto';
 export * from './delete-user.dto';
 export * from './dto';
 export * from './get-airdrop-balances.dto';
